Create a uWSGI postfork function registry. Also, start the tool conf watcher thread post-fork.
