+++ lib/galaxy/webapps/galaxy/buildapp.py
@@ -13,7 +13,6 @@ except:
 
 
 import galaxy.app
-from galaxy.config import process_is_uwsgi
 import galaxy.model
 import galaxy.model.mapping
 import galaxy.datatypes.registry
@@ -22,21 +21,13 @@ import galaxy.web.framework.webapp
 from galaxy.webapps.util import build_template_error_formatters
 from galaxy import util
 from galaxy.util import asbool
+from galaxy.util.postfork import process_is_uwsgi, register_postfork_function
 from galaxy.util.properties import load_app_properties
 
 from paste import httpexceptions
 
 import logging
 log = logging.getLogger( __name__ )
-
-try:
-    from uwsgidecorators import postfork
-except:
-    # TODO:  Make this function more like flask's @before_first_request w/
-    # registered methods etc.
-    def pf_dec(func):
-        return func
-    postfork = pf_dec
 
 
 class GalaxyWebApplication( galaxy.web.framework.webapp.WebApplication ):
@@ -135,8 +126,7 @@ def paste_app_factory( global_conf, **kwargs ):
     except:
         log.exception("Unable to dispose of pooled toolshed install model database connections.")
 
-    if not process_is_uwsgi:
-        postfork_setup()
+    register_postfork_function(postfork_setup)
 
     # Return
     return webapp
@@ -158,7 +148,6 @@ def uwsgi_app_factory():
     return app_factory(global_conf, **kwargs)
 
 
-@postfork
 def postfork_setup():
     from galaxy.app import app
     if process_is_uwsgi:
