+++ lib/galaxy/webapps/tool_shed/buildapp.py
@@ -15,7 +15,7 @@ import galaxy.web.framework.webapp
 from galaxy.webapps.util import build_template_error_formatters
 from galaxy.webapps.tool_shed.framework.middleware import hg
 from galaxy import util
-from galaxy.config import process_is_uwsgi
+from galaxy.util.postfork import process_is_uwsgi
 from galaxy.util.properties import load_app_properties
 
 log = logging.getLogger( __name__ )
